# perf(compiler): switch to FxHashMap/FxHashSet