Switches to a more performant hashing used by rustc (variant of FNV). Overall compilation is ~25% faster, but mono is ~50% faster. Parsing still dominates.

Mini-bench of 100 runs (compiling stdlib tests)

## Old timings

| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 140 | 126 | 199 | 144.65 | 16.12 |
| Pass1 | 13 | 11 | 20 | 13.49 | 2.11 |
| Ast | 123 | 111 | 193 | 130.07 | 16.40 |
| Mono | 148 | 126 | 202 | 152.64 | 17.35 |
| Optimizations | 8 | 7 | 11 | 8.03 | 0.79 |
| Codegen | 45 | 39 | 71 | 46.50 | 6.77 |
| TOTAL | 503 | 439 | 595 | 504.68 | 36.45 |

## New timings

| Stage | Median time (ms) | Min time (ms) | Max time (ms) | Average time (ms) | Standard deviation (ms) |
| --- | --: | --: | --: | --: | --: |
| Init | 0 | 0 | 0 | 0.00 | 0.00 |
| Parse | 134 | 121 | 199 | 140.00 | 17.37 |
| Pass1 | 12 | 11 | 19 | 12.93 | 2.18 |
| Ast | 97 | 87 | 152 | 101.86 | 13.29 |
| Mono | 102 | 90 | 147 | 105.02 | 11.19 |
| Optimizations | 5 | 4 | 8 | 5.21 | 0.78 |
| Codegen | 32 | 28 | 54 | 33.46 | 5.32 |
| TOTAL | 401 | 355 | 500 | 407.37 | 30.62 |